searchRepositoriesでひとまとめ担っていた処理を各関数に分けた
